pyproject.toml
[project]
name = "shamela-to-kindle"
version = "0.1.0"
description = "Convert books from shamela.ws to Kindle-friendly EPUB (RTL)"
authors = [{ name = "Shamela Tools" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = []

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0",
  "black>=24.0",
]
cover = [
  "pillow>=10.0",
]

[project.scripts]
# Route console entry through the package (library-first).
shamela-to-epub = "shamela_books.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["shamela_books*"]
